refactored profile actions for simplicity
