+++ client/src/components/Header/Header.scss
@@ -33,6 +33,7 @@
         font-weight: bold;
         margin: 0px;
         padding: 0px;
+        cursor: default;
     }
 
     p {
@@ -40,5 +41,6 @@
         line-height: $subtitle_lineheight_size;
         margin: 0px;
         padding: 0px;
+        cursor: default;
     }
 }
